+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -467,6 +467,17 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             menuIndicatorForegroundSchemeColor:
                 controller.menuIndicatorForegroundSchemeColor,
             menuIndicatorRadius: controller.menuIndicatorRadius,
+            // SearchView and SearchBar settings.
+            searchBarBackgroundSchemeColor:
+                controller.searchBackgroundSchemeColor,
+            searchViewBackgroundSchemeColor:
+                controller.searchBackgroundSchemeColor,
+            searchBarElevation: controller.searchElevation,
+            searchViewElevation: controller.searchElevation,
+            searchBarRadius: controller.searchRadius,
+            searchViewRadius: controller.searchRadius,
+            searchViewHeaderHeight: controller.searchViewHeaderHeight,
+            searchUseGlobalShape: controller.searchUseGlobalShape,
             // NavigationBar settings.
             navigationBarSelectedLabelSchemeColor:
                 controller.navBarSelectedLabelSchemeColor,